+++ lib/src/codecs/codecs.dart
@@ -28,3 +28,5 @@ export 'any_uri/any_uri_codec.dart';
 export 'decimal/decimal_codec.dart';
 export 'double/double_codec.dart';
 export 'float/float_codec.dart';
+export 'integer/integer_codec.dart';
+export 'unsigned_int/unsigned_int_codec.dart';