+++ lib/src/codecs/integer/integer_encoder.dart
@@ -1,0 +1,8 @@
+import 'dart:convert';
+
+class XsdIntegerEncoder extends Converter<BigInt, String> {
+  const XsdIntegerEncoder();
+
+  @override
+  String convert(BigInt input) => input.toString();
+}